app_python/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "app to show the current MSK time"
authors = ["Danila Danko <br4ch1st0chr0n3@gmail.com>"]

[tool.poetry.dependencies]
uvicorn = "^0.18.3"
fastapi = "^0.85.0"
python-dotenv = "^0.21.0"
python = "^3.10"
Jinja2 = "^3.1.2"
click = "^8.1.3"
bs4 = "^0.0.1"
pytest = "^7.1.3"

[tool.poetry.dev-dependencies]
uvicorn = "^0.18.3"
manage-fastapi = "^1.1.1"
pylint = "^2.15.3"

[tool.poetry.scripts]
app = 'app.main:main'

[tool.poetry.group.dev.dependencies]
colorama = "^0.4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# [tool.black]
# # https://github.com/psf/black
# line-length = 88
# target_version = ['py310']
# exclude = '''
# (
#   /(
#     \.git
#     | \.mypy_cache
#     | \.pytest_cache
#     | htmlcov
#     | venv
#     | .venv
#   )/
# )
# '''

# [tool.isort]
# # https://github.com/timothycrosley/isort
# # https://github.com/timothycrosley/isort/wiki/isort-Settings
# line_length = 88
# indent = '    '
# multi_line_output = 3
# include_trailing_comma = true
# force_grid_wrap = 0


# [tool.coverage]
# # https://github.com/nedbat/coveragepy
# [tool.coverage.run]

# source = ["app"]
# branch = true
# omit = ['']

# [tool.coverage.report]
# exclude_lines = ["# noqa", "raise NotImplementedError"]
